Fixes, changes and updates
.: Fixed: order in start-up sequence
.: Changed: use uint32 for upload and download speed limits
.: Fixed: possible memory leak in collections
.: Removed: MobileMule code
.: AsyncSocketEx code: fixed regressions, updated old OnLayerCallback methods
.: Added: mp4, mkv and ogg quick verification (thanks to Enig123)
.: Fixed: regression in Web Interface
.: Fixed: Shared Files page should always have a '+' symbol for "All Directories"
.: Changed: allow mediainfo.dll up to 18.05.* version
.: Fixed: string type inconsistency in KAD tags (in non-Unicode builds only)

+++ Scheduler.h
@@ -67,8 +67,8 @@ public:
 	void	RestoreOriginals();
 	void	ActivateSchedule(INT_PTR index,bool makedefault=false);
 
-	uint16	original_upload;
-	uint16	original_download;
+	uint32	original_upload;
+	uint32	original_download;
 	UINT	original_connections;
 	UINT	original_cons5s;
 	UINT	original_sources;
